Organize bulk operations and results into their own directories
