+++ lib/mongo/operation/write/bulk.rb
@@ -12,12 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'mongo/operation/write/bulk'
-require 'mongo/operation/write/delete'
-require 'mongo/operation/write/insert'
-require 'mongo/operation/write/update'
-require 'mongo/operation/write/create_index'
-require 'mongo/operation/write/drop_index'
-require 'mongo/operation/write/create_user'
-require 'mongo/operation/write/remove_user'
-require 'mongo/operation/write/command'
+require 'mongo/operation/write/bulk/bulk_delete'
+require 'mongo/operation/write/bulk/bulk_insert'
+require 'mongo/operation/write/bulk/bulk_update'
