+++ lib/mongo/operation/write/bulk/bulk_mergable.rb
