+++ lib/mongo/operation/write/bulk/legacy_bulk_mergable.rb
